tests/wpt/tests/html/semantics/document-metadata/the-base-element/base_about_blank.html

<!doctype html>
<meta charset=utf-8>
<title>base element in about:blank document should resolve against its fallback base URI</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<iframe></iframe>
<script>
var t = async_test();
addEventListener("load", t.step_func_done(function() {
  var doc = frames[0].document;
  var b = doc.createElement("base");
  b.setAttribute("href", "test");
  var newBaseValue = location.href.replace(/\/[^/]*$/, "/") + "test";
  assert_equals(b.href, newBaseValue);
  assert_equals(doc.baseURI, location.href);
  doc.head.appendChild(b);
  assert_equals(doc.baseURI, newBaseValue);
}));
</script>
