Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/semantics/document-metadata/the-base-element/base_about_blank-expected.txt



FAIL base element in about:blank document should resolve against its fallback base URI assert_equals: expected "http://localhost:8800/html/semantics/document-metadata/the-base-element/test" but got ""
